# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sosyal Medya API",
        "description": "backend",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Tüm postlar (kronolojik)",
                "operationId": "3fda362e2e082b5357ac416e7767e435",
                "responses": {
                    "200": {
                        "description": "Post listesi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Yeni gönderi oluştur (metin + birden fazla fotoğraf/video opsiyonel + konum)",
                "operationId": "dcb66200c3132996151884ac232cf439",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "caption"
                                ],
                                "properties": {
                                    "caption": {
                                        "type": "string",
                                        "example": "Tatil fotoğrafları!"
                                    },
                                    "media[]": {
                                        "description": "Birden fazla fotoğraf veya video (max 10 tane)",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Antalya, Türkiye"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post oluşturuldu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "caption": {
                                            "type": "string"
                                        },
                                        "media": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "media_urls": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/posts/{post}": {
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Post düzenle (sadece sahibi)",
                "operationId": "fc17b67462ad71a7f4b287520462fd02",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "caption": {
                                        "type": "string",
                                        "example": "Güncellenmiş başlık"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "İstanbul, Türkiye"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post güncellendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Post başarıyla güncellendi"
                                        },
                                        "post": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Yetkisiz işlem"
                    },
                    "404": {
                        "description": "Post bulunamadı"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Post sil (sadece sahibi)",
                "operationId": "f0dcb391948b9589ee732a26a68e17c7",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post başarıyla silindi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Post başarıyla silindi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Yetkisiz işlem"
                    },
                    "404": {
                        "description": "Post bulunamadı"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/feed": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Ana sayfa (takip edilenler + kendi postlar)",
                "operationId": "00be9f462a59b23e3ddfa04612b56a03",
                "responses": {
                    "200": {
                        "description": "Feed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/my-posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Kendi postlarımı listele",
                "operationId": "b1b8c2c6c33c80f172a3c5694b386acd",
                "responses": {
                    "200": {
                        "description": "Kendi postlar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/posts/{post}/archive": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Gönderiyi arşivle",
                "operationId": "4bed4a3b74f94f54a6dfcc8c8de8aef8",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gönderi başarıyla arşivlendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Gönderi başarıyla arşivlendi"
                                        },
                                        "post": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Yetkisiz işlem"
                    },
                    "404": {
                        "description": "Post bulunamadı"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/posts/{post}/unarchive": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Gönderiyi arşivden çıkar",
                "operationId": "df30fb8401c23a83dc6f8391ef13cad4",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gönderi başarıyla arşivden çıkarıldı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Gönderi başarıyla arşivden çıkarıldı"
                                        },
                                        "post": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Yetkisiz işlem"
                    },
                    "404": {
                        "description": "Post bulunamadı"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": {
        "1": {
            "sanctum": []
        }
    }
}